Add command pipe support through :pipe option
Allows for specifying the :pipe option, which can be used to pass nested
pipes to the command that is parsed in the same fashion as the base
command.  This also allows for piped commands to go through the same
sanitation helpers in the CommandLineBuilder as the base command, though
passing an unfiltered string is also supported.